[Feat] Rename package name (async_client_decorator -> async_client)
