+++ example/single_session.py
@@ -3,7 +3,7 @@ from typing import NamedTuple
 
 import aiohttp
 
-from async_client_decorator import request, Session, Query
+from async_client import request, Session, Query
 
 loop = asyncio.get_event_loop()
